fixup! Added Timer [Not Working]
I just wasted one whole night in debuging the why timer is not working.
You would't belive that i was so much feed up that i was thinking of
dumping this idea i was trying to create a new brach and start over
thank god i don't know that much of git other wise right now i might we
wasting my time in rewriting it.Do you want to know the solution IRQ was
genrating because it cli i jst have to sti. In the relam of fixing it i
have made to much mess i have to clear it now.
Update: clean all the mess.